Normalise behaviour of nodeRange and nodeRangeM
The default method implementation throws an error on empty
graphs (albeit one for being unable to find the minimum of an empty
list).

This makes the error more obvious, and also make the two Graph instances
have the same error (rather than returning a result that is identical
that from `nodeRange (mkGraph [(0,"only node")] [])`).