Make equal work for graphs with multiple edges
With the previous version, as there is no Ord constraint on edge labels
it is not possible guarantee that a list of multiple edges would equal
each other as there is no way to tell how they would be ordered in the
list produced by labEdges.

A simpler variant would have been to use eqList on labEdges directly;
however, this approach at least provides *some* performance
improvements (as the usage of \\ in eqList is O(n^2), so using it on
smaller sub-lists is preferable).

Note that this also assumes that a graph instance is sane enough to not
have repeated Node values in the output of labNodes; otherwise, the same
problem would occur.

Also remove the custom comparison operator for slabNodes.