Clarify how edges are removed from the graph.
Previously, delLEdge deleted *all* edges with that same label, which
might not be what was expected in the case of multiple edges with the
same label (and doesn't match the specification as per the documentation
for it).

Add 'delAllLEdges' to use this behaviour, and re-define delLEdge to only
delete one such labelled edge.