Better definitions of mkGraph
These definitions assume that the list of nodes has no repeated node
values, but so did the originals: they ultimately both used `&` which -
for both Tree and PatriciaTree - throws an error if a node already
exists.

Instead, create the underlying Map/IntMap directly containing the nodes,
and then add the edges (rather than continually adding one node at a
time).

With the new version of insEdges, there's no need for a custom usage of
foldl'.