Generate sample data more specifically
This means that not all properties are just simple Bool-based
expressions, but they're also more sensible now by being able to use the
smarter generators specifically.

As part of this, tests should *not* be internally exhaustive (otherwise
they become too computationally expensive): instead, generate sample
data based upon the actual graph (which is not possible to do with just
Bool-based properties, and is the main reason why Gen-based properties
are now being used).  For example, see the new definition of
`valid_match`.