Use differenceWith to implement (&) and match
Use `differenceWith` to implement `(&)` and `match` for
`Data.Graph.Inductive.PatriciaTree`. This is much better than
making many changes one by one (when the node is linked to
many others), but it's not so great when making
just a couple. It looks like the optimal threshold is somewhere
around 5, so that's where I've set it for `(&)`. I haven't
set a threshold for `match`, because it's a little harder to
see how to do that.

One other thing we may want to consider is batching up the
predecessor and successor modifications into a single map
to use with `differenceWith`. I don't have a clear sense of
whether that will be better or worse.