Fix implementation of rc in TransClos.
As observed in issue #86,
the implementation of the function rc
in the module Data.Graph.Inductive.Query.TransClos.hs
computes the graph with only the self-loops,
rather than the reflexive closure of a graph.

This change
conserves the existing edges,
although it drops their labels in the process
to satisfy the desired type of the function.