Updating upper bound versions of dependencies.
fgl-arbitrary is a dependency of graphviz, which would be great to have included in recent and current LTS versions in stackage. In order to achieve this, fgl-arbitrary would be one of the component to be updated in hackage.

A new version would be nice to have on Hackage, but updating the version bounds in the currently available version is also a working solution which I have just tested.