Remove MonadFail from being a superclass of GraphM.
This adds support for GHC 9.4, where ST is no longer an instance of
MonadFail.  The two instances of GraphM that are predefined in fgl are
ST and IO, and these always defined 'fail = error', so for these there
is no change in behaviour.  This may be a change in behaviour for
other instances of GraphM.  Users can work around this by defining
their own methods for matchAnyM and nodeRangeM that use 'fail' instead
of 'error'.